requirements.txt
Django==1.11.4
django-filter==1.0.4
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
drfdocs==0.0.11
olefile==0.44
Pillow==4.2.1
PyJWT==1.5.3
pytz==2017.2